Renamed config and discovery packages to internal in the HDPowerViewHub binding. See openhab/static-code-analysis#100
