+++ addons/binding/org.openhab.binding.hdpowerview/OSGI-INF/HDPowerViewDiscovery.xml
@@ -10,7 +10,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" modified="modified" name="org.openhab.binding.hdpowerview.discovery.HDPowerViewHubDiscoveryService">
-   <implementation class="org.openhab.binding.hdpowerview.discovery.HDPowerViewHubDiscoveryService"/>
+   <implementation class="org.openhab.binding.hdpowerview.internal.discovery.HDPowerViewHubDiscoveryService"/>
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
    </service>
